@@ -1,6 +1,6 @@
 plugins {
     id "org.jetbrains.gradle.plugin.idea-ext" version "1.1.7"
-    id 'io.github.CDAGaming.cursegradle' version '1.6.+'
+    id 'net.darkhax.curseforgegradle' version '1.1.15'
     id 'com.modrinth.minotaur' version '2.8.+'
 }
 
@@ -9,10 +9,10 @@ subprojects {
     apply plugin: 'eclipse'
     apply plugin: 'idea'
     apply plugin: 'maven-publish'
-    apply plugin: 'io.github.CDAGaming.cursegradle'
+    apply plugin: 'net.darkhax.curseforgegradle'
     apply plugin: 'com.modrinth.minotaur'
 
-    java.toolchain.languageVersion = JavaLanguageVersion.of(17)
+    java.toolchain.languageVersion = JavaLanguageVersion.of(21)
     java.withSourcesJar()
 
     ext.projectProperties = [
@@ -78,7 +78,7 @@ subprojects {
             inputs.properties project.ext.projectProperties
             project.ext.projectProperties.put 'project', project
 
-            filesMatching(['META-INF/mods.toml', 'fabric.mod.json', 'pack.mcmeta']) {
+            filesMatching(['META-INF/neoforge.mods.toml', 'fabric.mod.json', 'pack.mcmeta']) {
                 expand project.ext.projectProperties
             }
 
@@ -89,7 +89,7 @@ subprojects {
 
     tasks.withType(JavaCompile).configureEach {
         it.options.encoding = 'UTF-8'
-        it.options.release = 17
+        it.options.release = 21
     }
 
     tasks.withType(GenerateModuleMetadata) {
