@@ -9,42 +9,34 @@ import net.tslat.effectslib.api.ExtendedMobEffect;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Slice;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(GameRenderer.class)
 public class GameRendererMixin {
-	@Shadow @Final Minecraft minecraft;
+	@Shadow
+	@Final
+	Minecraft minecraft;
 
 	@Inject(
 			method = "render",
-			slice = @Slice(
-					from = @At(
-							value = "INVOKE",
-							target = "Lcom/mojang/blaze3d/systems/RenderSystem;applyModelViewMatrix()V"
-					),
-					to = @At(
-							value = "INVOKE",
-							target = "Lnet/minecraft/client/renderer/GameRenderer;renderConfusionOverlay(Lnet/minecraft/client/gui/GuiGraphics;F)V"
-					)
-			),
 			at = @At(
-					value = "INVOKE_ASSIGN",
-					target = "Ljava/lang/Double;floatValue()F",
-					shift = At.Shift.AFTER
+					value = "INVOKE",
+					target = "Lnet/minecraft/client/Options;screenEffectScale()Lnet/minecraft/client/OptionInstance;"
 			)
 	)
-	private void renderEffectOverlays(float partialTicks, long nanoTime, boolean renderLevel, CallbackInfo callback) {
-		doExtendedEffectRenders(this.minecraft.player, partialTicks);
+	private void tel$renderEffectOverlays(float partialTicks, long nanoTime, boolean renderLevel, CallbackInfo callback) {
+		tel$doExtendedEffectRenders(this.minecraft.player, partialTicks);
 	}
 
-	private void doExtendedEffectRenders(LocalPlayer player, float partialTicks) {
+	@Unique
+	private void tel$doExtendedEffectRenders(LocalPlayer player, float partialTicks) {
 		PoseStack poseStack = new PoseStack();
 
 		for (MobEffectInstance instance : player.getActiveEffects()) {
-			if (instance.getEffect() instanceof ExtendedMobEffect extendedMobEffect && extendedMobEffect.getOverlayRenderer() != null)
+			if (instance.getEffect().value() instanceof ExtendedMobEffect extendedMobEffect && extendedMobEffect.getOverlayRenderer() != null)
 				extendedMobEffect.getOverlayRenderer().render(poseStack, partialTicks, instance);
 		}
 	}
