@@ -1,5 +1,6 @@
 package net.tslat.effectslib.mixin.common;
 
+import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.inventory.EnchantmentMenu;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentInstance;
@@ -15,7 +16,7 @@ import java.util.List;
 @Mixin(EnchantmentMenu.class)
 public class EnchantmentMenuMixin {
 	@Inject(method = "getEnchantmentList", at = @At("TAIL"), locals = LocalCapture.CAPTURE_FAILSOFT)
-	private void filterExtendedEnchantments(ItemStack stack, int enchantSlot, int level, CallbackInfoReturnable<List<EnchantmentInstance>> callback, List<EnchantmentInstance> enchants) {
+	private void filterExtendedEnchantments(FeatureFlagSet featureFlags, ItemStack stack, int enchantSlot, int level, CallbackInfoReturnable<List<EnchantmentInstance>> callback, List<EnchantmentInstance> enchants) {
 		enchants.removeIf(instance -> instance.enchantment instanceof ExtendedEnchantment extendedEnchant && !extendedEnchant.canEnchant(stack, ExtendedEnchantment.ENCHANTING_TABLE));
 	}
 }
