@@ -2,33 +2,38 @@ package net.tslat.effectslib.mixin.common;
 
 import net.minecraft.world.inventory.GrindstoneMenu;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.ItemEnchantments;
 import net.tslat.effectslib.api.ExtendedEnchantment;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import java.util.Map;
 
 @Mixin(GrindstoneMenu.class)
 public class GrindstoneMenuMixin {
 	@Unique
-	private static ItemStack tslatEffectsLib$capturedStack;
+	private static ItemStack tel$capturedStack;
 
-	@Inject(method = "removeNonCurses", at = @At("HEAD"))
-	public void captureTelStack(ItemStack stack, int damage, int count, CallbackInfoReturnable<ItemStack> cir) {
-		tslatEffectsLib$capturedStack = stack;
+	@Inject(method = "removeNonCursesFrom", at = @At("HEAD"))
+	public void tel$captureStack(ItemStack stack, CallbackInfoReturnable<ItemStack> callback) {
+		tel$capturedStack = stack;
 	}
 
-	@Inject(method = {"lambda$removeNonCurses$0", "method_16694"}, at = @At(value = "HEAD", remap = false), cancellable = true)
-	private static void canEnchantExtended(Map.Entry<Enchantment, Integer> entry, CallbackInfoReturnable<Boolean> cir) {
-		if (entry.getKey() instanceof ExtendedEnchantment extendedEnchantment) {
-			Boolean shouldRemove = extendedEnchantment.shouldGrindstoneRemove(tslatEffectsLib$capturedStack);
+	@Inject(method = {"lambda$removeNonCursesFrom$2", "method_16694"}, at = @At(value = "HEAD", remap = false), cancellable = true)
+	private static void tel$checkRemoveCurses(ItemEnchantments.Mutable enchantments, CallbackInfo callback) {
+		enchantments.removeIf(enchant -> {
+			if (enchant instanceof ExtendedEnchantment extendedEnchantment) {
+				Boolean shouldRemove = extendedEnchantment.shouldGrindstoneRemove(tel$capturedStack);
 
-			if (shouldRemove != null)
-				cir.setReturnValue(!shouldRemove);
-		}
+				if (shouldRemove != null)
+					return shouldRemove;
+			}
+
+			return !enchant.value().isCurse();
+		});
+
+		callback.cancel();
 	}
 }
