@@ -3,13 +3,12 @@ package net.tslat.effectslib;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
-import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
+import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.tslat.effectslib.networking.packet.MultiloaderPacket;
-
-import java.util.Locale;
-import java.util.function.Function;
+import org.jetbrains.annotations.ApiStatus;
 
 public class TELFabricClient implements ClientModInitializer {
     @Override
@@ -18,14 +17,12 @@ public class TELFabricClient implements ClientModInitializer {
     }
 
     public static void sendPacketToServer(MultiloaderPacket packet) {
-        FriendlyByteBuf buffer = PacketByteBufs.create();
-
-        packet.write(buffer);
-
-        ClientPlayNetworking.send(packet.id(), buffer);
+        ClientPlayNetworking.send(packet);
     }
 
-    public static <P extends MultiloaderPacket> void registerPacket(Class<P> messageType, Function<FriendlyByteBuf, P> decoder) {
-        ClientPlayNetworking.registerGlobalReceiver(new ResourceLocation(TELConstants.MOD_ID, messageType.getName().toLowerCase(Locale.ROOT)), (client, handler, buf, responseSender) -> decoder.apply(buf).receiveMessage(client.player, client::execute));
+    @ApiStatus.Internal
+    public static <B extends FriendlyByteBuf, P extends MultiloaderPacket> void registerPacket(CustomPacketPayload.Type<P> packetType, StreamCodec<B, P> codec) {
+        PayloadTypeRegistry.playS2C().register(packetType, (StreamCodec<FriendlyByteBuf, P>)codec);
+        ClientPlayNetworking.registerGlobalReceiver(packetType, (packet, context) -> packet.receiveMessage(context.player(), context.client()::execute));
     }
 }
