@@ -2,13 +2,13 @@ package net.tslat.effectslib;
 
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.Mod;
-import net.neoforged.neoforge.event.TickEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.client.event.ClientTickEvent;
 
-@Mod.EventBusSubscriber(value = Dist.CLIENT)
+@EventBusSubscriber(value = Dist.CLIENT)
 public final class TELNeoForgeClient {
     @SubscribeEvent
-    public static void clientTick(final TickEvent.ClientTickEvent ev) {
+    public static void clientTick(final ClientTickEvent.Pre ev) {
         TELClient.tickParticleTransitions();
     }
 }
